Updated the product price to be a decimal value.

+++ database/migrations/2018_02_22_151247_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->double('total_price');
+            $table->decimal('total_price',10,2);
             $table->timestamps();
         });
